requirements.txt
flask
python-dotenv
google-generativeai>=0.4.1
flask-cors
uvicorn